[core] add a removeAllListeners() call to EventDispatchers (incl. Sprites) (Fix #378) Avoid integer rounding in some font related computations

+++ 2dsg/ttbmfont.cpp
@@ -826,7 +826,7 @@ float TTBMFont::getAdvanceX(const char *text, float letterSpacing, int size) {
 
 		x += textureGlyph.advancex >> 6;
 
-		x += (int) (letterSpacing / sizescalex_);
+		x += (letterSpacing / sizescalex_);
 	}
 
 	x += kerning(prev, wtext[size]) >> 6;
